+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## 1.2.23
+Made a new option for snowdepth. Azimuth dependent bare soil correction is now
+the default.  
+
 Changed RHdot calculation in subdaily to default behavior.
 
 ## 1.2.22